Revert "Code Refactoring"
This reverts commit 9ae530fefe2c11e9a9b4e882abb185c490abd519.
