--- shared.py
@@ -1,5 +1,0 @@
-import os
-
-def clear_screen():
-    # Clear the console screen
-    os.system('cls' if os.name == 'nt' else 'clear')
